Check whether file exists and add update flag
In load and subset simulation data function

+++ simulation_study/simulation_stan_analysis/subset_simulated_data.R
@@ -17,40 +17,45 @@ load_and_subset_sim_data = function(min_entry_size,
                                     sd,
                                     target_num_rows_data,
                                     num_line_data,
-                                    num_container_data) {
+                                    num_container_data,
+                                    update = FALSE) {
   
-  # open full data frame
-  data_filename <- gen_filename_simulate_data(min_entry_size,max_entry_size,sd,target_num_rows_data)
-  load_sim_data <- read.csv(data_filename)
-  
-  # check that the number of rows for both line and contianer don't exceed the max
-  max_line = length(load_sim_data %>% filter(Mode == 'Line') %>% pull(Mode))
-  max_container <- length(load_sim_data %>% filter(Mode == 'Container') %>% pull(Mode))
-  
-  if(num_line_data > max_line){
-    message(paste0("Input number of lines ", num_line_data, " > max=", max_line, ". Setting to max..." ))
-    num_line_data <- max_line
-  }
-  if(num_container_data > max_container){
-    message(paste0("Input number of containers ", num_container_data, " > max=", max_container, ". Setting to max..." ))
-    num_container_data <- max_container
-  }
-  
-  # extract line_rows of line data and container_rows of container data
-  line_data <- load_sim_data %>% filter(Mode == 'Line') %>% head(num_line_data)
-  container_data <- load_sim_data %>% filter(Mode == 'Container') %>% head(num_container_data)
-  load_sim_data <- rbind(line_data, container_data)
-
-  #load_sim_data <- load_sim_data %>% mutate(Country=1)
-  #load_sim_data <- load_sim_data %>% mutate(Entry=1)
-  
-  fit_summary_df <- run_stan_simulated_data(load_sim_data,
-                                            data_filename,
-                                            entry_correlation_flag = FALSE
-  )
-
   output_filename = gen_filename_stanfit_simulated_data(min_entry_size, max_entry_size,sd,num_line_data, num_container_data)
   
-  saveRDS(fit_summary_df, output_filename)
+  if(!file.exists(output_filename) | update){
+    # open full data frame
+    data_filename <- gen_filename_simulate_data(min_entry_size,max_entry_size,sd,target_num_rows_data)
+    load_sim_data <- read.csv(data_filename)
+    
+    # check that the number of rows for both line and container don't exceed the max
+    max_line = length(load_sim_data %>% filter(Mode == 'Line') %>% pull(Mode))
+    max_container <- length(load_sim_data %>% filter(Mode == 'Container') %>% pull(Mode))
+    
+    if(num_line_data > max_line){
+      message(paste0("Input number of lines ", num_line_data, " > max=", max_line, ". Setting to max..." ))
+      num_line_data <- max_line
+    }
+    if(num_container_data > max_container){
+      message(paste0("Input number of containers ", num_container_data, " > max=", max_container, ". Setting to max..." ))
+      num_container_data <- max_container
+    }
+    output_filename = gen_filename_stanfit_simulated_data(min_entry_size, max_entry_size,sd,num_line_data, num_container_data)
+    
+    # extract line_rows of line data and container_rows of container data
+    line_data <- load_sim_data %>% filter(Mode == 'Line') %>% head(num_line_data)
+    container_data <- load_sim_data %>% filter(Mode == 'Container') %>% head(num_container_data)
+    load_sim_data <- rbind(line_data, container_data)
+    
+    #load_sim_data <- load_sim_data %>% mutate(Country=1)
+    #load_sim_data <- load_sim_data %>% mutate(Entry=1)
+    
+    fit_summary_df <- run_stan_simulated_data(load_sim_data,
+                                              data_filename,
+                                              entry_correlation_flag = FALSE
+    )
+    
+    saveRDS(fit_summary_df, output_filename)
+  }
+  
   output_filename
 }
